requirements.txt
# AdSavant Backend Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9

# ML Runtime
numpy<2
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
joblib>=1.3.0
pandas>=2.0.0

# CLIP
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.24.0

# NLP
textblob>=0.18.0
emoji>=2.14.0

# Image
Pillow>=10.0.0
